log_pipeline/requirements.txt
# Required for real Gemini API integration
google-generativeai==0.4.0
